+++ Md5Parser/Parser/Data/Mesh/Vertex.cs
@@ -2,7 +2,7 @@
 {
     public class Vertex
     {
-        public Vertex(int index, double texU, double texV, int weightIndex, int weightElem)
+        public Vertex(int index, float texU, float texV, int weightIndex, int weightElem)
         {
             Index = index;
             TexU = texU;
@@ -12,8 +12,8 @@
         }
 
         public int Index { get; private set; }
-        public double TexU { get; private set; }
-        public double TexV { get; private set; }
+        public float TexU { get; private set; }
+        public float TexV { get; private set; }
         public int WeightIndex { get; private set; }
         public int WeightElem { get; private set; }
     }
